Set env vars consistently in postgres config doc
YAML is flexible, so it allows a couple ways to specify env vars for docker images. This just picks one of the two ways, to be consistent.

Why? Because I accidentally combined both (`- PGUSER=ubuntu` and `PGUSER: ubuntu` -> `PGUSER=ubuntu`), with very confusing results, since this caused all my env vars to coalesce as one, and no error was raised.